+++ src/renderer/components/EditPage/LeftPanelTable.vue
@@ -2,9 +2,10 @@
   <div style="overflow:auto;">
     <table id="edit-leftpaneltable-table">
       <tr>
-        <th colspan="10" class="table-danger"> 选择编辑条目</th>
+        <th colspan="10" class="table-info"> {{fileName}}（共有{{fileLength}}条记录）</th>
       </tr>
-      <tr class="edit-leftpaneltable-tr" v-for="(data, index) in file" v-bind:key='index' v-on:click="loadDoc(data, index)" v-bind:class="{'table-danger':flag === index && index !== 0}">
+      <tr class="edit-leftpaneltable-tr" v-for="(data, index) in file" v-bind:key='index' v-on:click="loadDoc(data, index)" v-bind:class="{'table-warning':flag === index}">
+        <td> {{index + 1}} </td>
         <td v-if="lastNav !== '/edit' && index < 10" v-for="(field, index) in data" v-bind:key='index' v-on:click="onClickTd(data, index)" v-bind:class="{'table-danger':flagTd.find((n)=>n===index)}">{{data[index]}}</td>
         <td v-if="lastNav === '/edit'">{{data.substr(0, 100)}}</td>
       </tr>
@@ -20,30 +21,36 @@
           return this.$store.state.Edit.lastNav
         }
       },
+      fileName: {
+        get() {
+          return this.$store.state.Edit.fileName
+        }
+      },
+      fileLength: {
+        get() {
+          return this.$store.state.Edit.file.length
+        }
+      },
       file: {
         get() {
           let f = []
-          if (this.$store.state.Edit.lastNav === '/edit') {
-            let start = 0
-            let fileLen = this.$store.state.Edit.file.length;
-            if (fileLen > 99) {
-              if (this.$store.state.Edit.filePage > 0) {
-                start = 100 * this.$store.state.Edit.filePage
-                fileLen = start + 99
-              } else {
-                fileLen = 99
-              }
-            }
-            for (let i = start; i < fileLen; i += 1) {
-              f.push(this.$store.state.Edit.file[i])
-            }
-          } else {
-            const type = typeof this.$store.state.Edit.file[0]
-            if (type === 'object') {
-              f = this.$store.state.Edit.file
+          const file = this.$store.state.Edit.file
+          let start = 0
+          let fileLen = this.$store.state.Edit.file.length;
+          if (fileLen > 100) {
+            if (this.$store.state.Edit.filePage > 0) {
+              start = 100 * this.$store.state.Edit.filePage
+              fileLen = start + 100
             } else {
-              f = this.$store.state.Edit.file.map(n => n.split(','))
+              fileLen = 100
             }
+          }
+          for (let i = start; i < fileLen; i += 1) {
+            f.push(file[i])
+          }
+          const type = typeof file[0]
+          if (this.$store.state.Edit.lastNav !== '/edit' && type !== 'object') {
+            f = f.map(n => n.split(','))
           }
           return f
         }