+++ src/renderer/store/modules/Stat.js
@@ -219,7 +219,16 @@ const mutations = {
     })
   },
   STAT_SET_SERVER_TABLE(state, opt) {
-    console.log(opt)
+    // state.tableSel = opt.tableSel
+    // state.dimensionOrg = opt.dimensionOrg
+    // state.dimensionTime = opt.dimensionTime
+    // state.dimensionDrg = opt.dimensionDrg
+    state.notice = [
+      `病案总数：${opt.tableSel}`,
+      `机构总数：${opt.dimensionOrg}`,
+      `时间维度总数：${opt.dimensionTime}`,
+      `病种维度总数：${opt.dimensionDrg}`,
+    ]
     state.isServer = true
     state.serverTable = opt
   },
@@ -285,17 +294,18 @@ const mutations = {
     state.localTable = state.localTables[state.tablePage]
   },
   STAT_SET_COL_NUM(state, num) {
-    console.log(num)
-    state.colNum = num
-    const indexs = []
-    // const rangeArray = (start, end) => Array((end - start) + 1).map((v, i) => i + start)
-    for (let i = num - 10; i < num; i += 1) { indexs.push(i) }
-    state.haveRight = true
-    const table = []
-    state.localTables[state.tablePage].forEach((xs) => {
-      table.push(indexs.map(x => xs[x]))
-    })
-    state.localTable = table
+    if (state.tableType === 'local') {
+      state.colNum = num
+      const indexs = []
+      // const rangeArray = (start, end) => Array((end - start) + 1).map((v, i) => i + start)
+      for (let i = num - 10; i < num; i += 1) { indexs.push(i) }
+      state.haveRight = true
+      const table = []
+      state.localTables[state.tablePage].forEach((xs) => {
+        table.push(indexs.map(x => xs[x]))
+      })
+      state.localTable = table
+    }
   },
   // STAT_SET_TITLE_PAGE(state, num) {
   //   state.colNum = num